+++ src/algorithms/validation/check.cpp
@@ -7,10 +7,12 @@ All rights reserved (see LICENSE).
 
 */
 
+#include <algorithm>
 #include <mutex>
 #include <thread>
 #include <unordered_map>
 #include <unordered_set>
+#include <vector>
 
 #include "algorithms/validation/check.h"
 #include "algorithms/validation/choose_ETA.h"